Address CI comments
